Revert "Revert "pr-comment finalize が中間生成物の削除を Git にコミットしない""
