+++ src/commands/auto-issue-output.ts
@@ -3,6 +3,7 @@ import { mkdir, writeFile } from 'node:fs/promises';
 
 import { logger } from '../utils/logger.js';
 import { getErrorMessage } from '../utils/error-utils.js';
+import { createMockableFunction, type MockableFunction } from '../utils/mockable-function.js';
 import type {
   AutoIssueExecutionInfo,
   AutoIssueIssueEntry,
@@ -18,10 +19,10 @@ interface BuildPayloadParams {
 /**
  * auto-issueコマンドの結果からJSONペイロードを構築
  */
-export function buildAutoIssueJsonPayload({
+const buildAutoIssueJsonPayloadImpl = ({
   execution,
   results,
-}: BuildPayloadParams): AutoIssueJsonOutput {
+}: BuildPayloadParams): AutoIssueJsonOutput => {
   const issues: AutoIssueIssueEntry[] = results.map((result) => ({
     success: result.success,
     title: result.title ?? 'Unknown title',
@@ -43,15 +44,18 @@ export function buildAutoIssueJsonPayload({
     summary,
     issues,
   };
-}
+};
+
+export const buildAutoIssueJsonPayload: MockableFunction<typeof buildAutoIssueJsonPayloadImpl> =
+  createMockableFunction(buildAutoIssueJsonPayloadImpl);
 
 /**
  * JSONペイロードをファイルへ書き出し
  */
-export async function writeAutoIssueOutputFile(
+const writeAutoIssueOutputFileImpl = async (
   filePath: string,
   payload: AutoIssueJsonOutput,
-): Promise<void> {
+): Promise<void> => {
   const directory = path.dirname(filePath);
   const serialized = `${JSON.stringify(payload, null, 2)}\n`;
 
@@ -64,4 +68,7 @@ export async function writeAutoIssueOutputFile(
     logger.error(`Failed to write auto-issue results to ${filePath}: ${message}`);
     throw new Error(`Failed to write auto-issue output file (${filePath}): ${message}`);
   }
-}
+};
+
+export const writeAutoIssueOutputFile: MockableFunction<typeof writeAutoIssueOutputFileImpl> =
+  createMockableFunction(writeAutoIssueOutputFileImpl);
