+++ src/core/safety/instruction-validator.ts
@@ -46,7 +46,7 @@ export const DANGEROUS_PATTERNS: Record<string, string[]> = {
   ],
   configChanges: ['config', 'env', '環境変数', '設定', '.env', 'credentials', 'secret', 'token', 'password'],
   dbOperations: ['DROP', 'DELETE FROM', 'TRUNCATE', 'UPDATE', 'INSERT', 'ALTER', 'CREATE TABLE', 'migration'],
-  autoFix: ['修正して', '直して', 'fix it', 'auto fix', '自動修正', '書き換えて', 'refactor', '変更して', 'update'],
+  autoFix: ['自動修正', '修正して', '直して', 'fix it', 'auto fix', '書き換えて', 'refactor', '変更して', 'update'],
 };
 
 /**
@@ -92,27 +92,16 @@ export class InstructionValidator {
    */
   public static validate(instruction: string): ValidationResult {
     const normalized = instruction.trim();
-    const validatedAt = new Date().toISOString();
 
     if (!normalized) {
       return {
         isValid: true,
-        confidence: 'high',
-        reason: 'Empty instruction',
-        category: 'analysis',
-        validationMethod: 'pattern',
-        validatedAt,
       };
     }
 
     if (normalized.length > InstructionValidator.MAX_LENGTH) {
       return {
         isValid: false,
-        confidence: 'high',
-        reason: `Instruction exceeds maximum length (${InstructionValidator.MAX_LENGTH} characters)`,
-        category: 'execution',
-        validationMethod: 'pattern',
-        validatedAt,
         errorMessage: `Custom instruction exceeds maximum length (${InstructionValidator.MAX_LENGTH} characters)`,
       };
     }
@@ -121,11 +110,6 @@ export class InstructionValidator {
     if (detectedPattern) {
       return {
         isValid: false,
-        confidence: 'high',
-        reason: `Dangerous operation detected: "${detectedPattern}"`,
-        category: 'execution',
-        validationMethod: 'pattern',
-        validatedAt,
         errorMessage: `Invalid custom instruction: Dangerous operation detected: "${detectedPattern}". Custom instructions must be limited to analysis guidance only.`,
         detectedPattern,
       };
@@ -133,11 +117,6 @@ export class InstructionValidator {
 
     return {
       isValid: true,
-      confidence: 'high',
-      reason: 'No dangerous patterns detected',
-      category: 'analysis',
-      validationMethod: 'pattern',
-      validatedAt,
     };
   }
