+++ src/utils/logger.js
@@ -1,0 +1,1 @@
+export { logger } from './logger.ts';
